+++ app/Http/Controllers/Api/OfferController.php
@@ -6,7 +6,6 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Api\OfferStoreRequest;
 use App\Http\Requests\Api\OfferUpdateRequest;
 use App\Http\Requests\Api\OfferExternalUpdateRequest;
-use Illuminate\Http\Request;
 use App\Models\Offer;
 use App\Models\Institution;
 use App\Models\Competence;
@@ -15,14 +14,42 @@ use App\Models\Language;
 use App\Models\Huboffer;
 use App\Models\Offertype;
 use App\Models\Timestamp;
-
-use Illuminate\Support\Facades\Log;
-use Illuminate\Support\Facades\Auth;
+use Illuminate\Pagination\LengthAwarePaginator ;
 use Illuminate\Support\Facades\DB;
+
 
 
 class OfferController extends Controller
 {
+
+    /**
+     * Display a paginated listing of the resource.
+     *
+     * @param int $offerCount
+     * @return \Illuminate\Http\Response
+     */
+    public function paginatedOffers(int $offerCount)
+    {
+        $paginatedOffers = Offer::orderBy('id')->Paginate($offerCount);
+        $pageWithOffers = $this->restructurePaginateResponse([$this,'restructureJsonOutput'], $paginatedOffers);
+
+        return response()->json($pageWithOffers, 200);
+    }
+
+    /**
+     * Display a reduced paginated listing for tiles.
+     *
+     * @param int $offerCount
+     * @return \Illuminate\Http\Response
+     */
+    public function paginatedReducedOffers(int $offerCount)
+    {
+        $paginatedOffers = Offer::orderBy('id')->Paginate($offerCount);
+        $pageWithOffers = $this->restructurePaginateResponse([$this,'getReducedOfferJson'], $paginatedOffers);
+
+        return response()->json($pageWithOffers, 200);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -93,7 +120,7 @@ class OfferController extends Controller
      */
     public function show(Offer $offer)
     {
-        return response()->json($this->restructureJsonOutput($offer), 200);
+        return response()->json($this->restructureJsonOutput($offer, true), 200);
     }
 
     /**
@@ -200,12 +227,65 @@ class OfferController extends Controller
     }
 
     /**
+     * Remove "data" and "links" from page
+     *
+     * @param  \Illuminate\Pagination\LengthAwarePaginator $offer
+     * @return Array $page
+     */
+     private function restructurePageMetadata( LengthAwarePaginator $offer ) {
+        $page = $offer->toArray();
+        unset(
+            $page["data"],
+            $page["links"],
+        );
+        return $page;
+    }
+
+    /**
+     * Restructure paginated response
+     *
+     * @param  \Illuminate\Pagination\LengthAwarePaginator $paginated_offers
+     * @return Array array_merge($data,$page)
+     */
+    private function restructurePaginateResponse($resturctureOffer,LengthAwarePaginator $paginatedOffers)
+    {
+        $transformedOffers = $paginatedOffers
+        ->getCollection()
+        ->transform(function($offer) use($resturctureOffer) {
+            return $resturctureOffer($offer);
+        });
+
+        $data = array("data" => $transformedOffers );
+        $page = $this->restructurePageMetadata($paginatedOffers);
+        return array_merge($data,$page);
+    }
+
+    /**
+     * Returns a Sublist of offer-short List (filter on keyword)
+     *
+     * @param String $type
+     * @return \Illuminate\Http\Response
+     */
+    public function getOfferSubListWithKeyword( String $keyword ) {
+
+        $keyword= preg_replace('/\s+/', '', $keyword);
+        $offers = Offer::select('offers.*')
+            ->leftJoin('huboffers','offers.id', '=', 'huboffers.offer_id')
+            ->whereRaw("FIND_IN_SET(?, huboffers.keywords) > 0", $keyword)->get();
+        $output = array();
+        foreach ( $offers as $offer ) {
+            $output[] = $this->getReducedOfferJson($offer);
+        }
+        return response()->json($output, 200);
+    }
+
+    /**
      * Remodel the output of an offer
      *
      * @param  \App\Models\Offer $offer
      * @return Array $ret
      */
-    private function restructureJsonOutput( Offer $offer ) {
+    private function restructureJsonOutput( Offer $offer, Bool $showRelatedOffersDetail = false ) {
 
         $ret = $offer->toArray();
 
@@ -244,10 +324,18 @@ class OfferController extends Controller
             }
         }
 
+
         $ret["relatedOffers"] = [];
         if ( array_key_exists( "original_relations", $ret ) ) {
             foreach ( $ret["original_relations"] as $relation ) {
                 $ret["relatedOffers"][] = $relation["id"];
+                if ( $showRelatedOffersDetail ) {
+                    $ret["relatedOfferData"][] = array (
+                        "id" => $relation["id"],
+                        "title" => $relation["title"],
+                        "image" => $relation["image_path"]
+                    );
+                }
             }
         }
 
@@ -288,7 +376,8 @@ class OfferController extends Controller
             "institution_id" => $offer->institution_id,
             "offertype_id" => $offer->offertype_id,
             "language_id" => $offer->language_id,
-            "competences" => $compeptences
+            "competences" => $compeptences,
+            "keywords" =>  isset( $offer->hubOffer) ? $offer->hubOffer->keywords : null
         );
 
         return $ret;