+++ config/bot.php
@@ -1,7 +1,0 @@
-<?php
-
-return [
-    'static_stage_url' => 'https://devstatic.futureskills-sh.de/data/kjdsihadjcfkatbb/bot.json',
-    'static_production_url' => 'https://static.futureskills-sh.de/data/kjdsihadjcfkatbb/bot.json',
-    'uuid_cookie_name' => 'FS_WAC_USER_ID'
-];