+++ Changelog.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## [Released] v0.4
+
+## 2021-09-20/dk
+## Changed
+- Route offer/{id} outputs some additional information about related offers
+
+## 2021-09-16/kw
+## Added
+- added paginated_offers function in Offercontroller (pagination kursliste)
+- added paginated_reduced_offers function in Offercontroller
+- added testroute {api}/offer/paginated/{offerCount}
+- added testroute{api}/list/offer/short/paginated/{offerCount}
+
+## 2021-09-14/ml
+### added
+- added function to get Offers by Keyword
+
+## 2021-09-03/dk
+### Changed
+- Upgraded Laravel to 8
+- Upgraded Passport to 10.1
+- JWT-Handling changed to support newer required lcobucci/jwt package
+### Removed 
+- Bot-related functions
+
+## 2021-09-01/ml
+### Changed
+- offerlist contains keywords (huboffer)
+
 ## [Released v0.3.1]
 
 ## 2021-08-10/kw